Improve workspace testability
- Modularize large test suites and expand coverage for CLI, rendering, Mermaid, and example entry points.
- Inject filesystem, clock, terminal, and output boundaries for deterministic failure-path testing.
- Register explicit integration tests and document the new testability boundaries.
