+++ crates/ag-tui-text/src/mermaid_test/support_test.rs
@@ -1,0 +1,10 @@
+use crate::mermaid::MermaidDiagram;
+
+pub(super) fn diagram_text(diagram: &MermaidDiagram) -> String {
+    diagram
+        .lines
+        .iter()
+        .map(ToString::to_string)
+        .collect::<Vec<_>>()
+        .join("\n")
+}
