+++ crates/ag-xtask/Cargo.toml
@@ -3,6 +3,7 @@ name = "ag-xtask"
 version.workspace = true
 authors.workspace = true
 edition.workspace = true
+license.workspace = true
 description.workspace = true
 repository.workspace = true
 publish = false
@@ -13,6 +14,7 @@ tracing.workspace = true
 tracing-subscriber.workspace = true
 
 [dev-dependencies]
+mockall.workspace = true
 tempfile.workspace = true
 
 [lints]
