+++ crates/testty/examples/proof_pipeline.rs
@@ -7,8 +7,8 @@
 //!
 //! Run with: `cargo run --example proof_pipeline -p testty`
 
-#![allow(clippy::print_stdout, clippy::expect_used)]
-
+use std::error::Error;
+use std::io::{self, Write};
 use std::path::{Path, PathBuf};
 
 use testty::frame::TerminalFrame;
@@ -19,9 +19,15 @@ use testty::proof::junit::JunitBackend;
 use testty::proof::report::ProofReport;
 use testty::proof::strip::ScreenshotStripBackend;
 
-fn main() {
-    let output_path = output_dir();
-    std::fs::create_dir_all(&output_path).expect("failed to create output dir");
+fn main() -> Result<(), Box<dyn Error>> {
+    run(
+        &output_dir(std::env::args().skip(1)),
+        &mut io::stdout().lock(),
+    )
+}
+
+fn run(output_path: &Path, output: &mut impl Write) -> Result<(), Box<dyn Error>> {
+    std::fs::create_dir_all(output_path)?;
 
     // Simulate a three-step user journey through a TUI application.
     let frame_startup = TerminalFrame::new(
@@ -69,69 +75,70 @@ fn main() {
     );
 
     // Render through all five backends.
-    println!("=== Testty Proof Pipeline Showcase ===\n");
+    writeln!(output, "=== Testty Proof Pipeline Showcase ===\n")?;
 
     // 1. Frame-text backend (annotated plain text).
     let text_path = output_path.join("proof.txt");
-    report
-        .save(&FrameTextBackend, &text_path)
-        .expect("frame-text render failed");
-    println!("Frame-text proof written to: {}", text_path.display());
-    print_file_preview(&text_path, 20);
+    report.save(&FrameTextBackend, &text_path)?;
+    writeln!(
+        output,
+        "Frame-text proof written to: {}",
+        text_path.display()
+    )?;
+    print_file_preview(&text_path, 20, output)?;
 
     // 2. Screenshot strip backend (vertical PNG).
     let strip_path = output_path.join("proof_strip.png");
-    report
-        .save(&ScreenshotStripBackend, &strip_path)
-        .expect("strip render failed");
+    report.save(&ScreenshotStripBackend, &strip_path)?;
     let strip_size = std::fs::metadata(&strip_path).map_or(0, |metadata| metadata.len());
-    println!(
+    writeln!(
+        output,
         "\nPNG strip written to: {} ({} bytes)",
         strip_path.display(),
         strip_size
-    );
+    )?;
 
     // 3. Animated GIF backend.
     let gif_path = output_path.join("proof.gif");
     let gif_backend = GifBackend::default();
-    report
-        .save(&gif_backend, &gif_path)
-        .expect("GIF render failed");
+    report.save(&gif_backend, &gif_path)?;
     let gif_size = std::fs::metadata(&gif_path).map_or(0, |metadata| metadata.len());
-    println!(
+    writeln!(
+        output,
         "Animated GIF written to: {} ({} bytes)",
         gif_path.display(),
         gif_size
-    );
+    )?;
 
     // 4. Self-contained HTML backend.
     let html_path = output_path.join("proof.html");
-    report
-        .save(&HtmlBackend, &html_path)
-        .expect("HTML render failed");
+    report.save(&HtmlBackend, &html_path)?;
     let html_size = std::fs::metadata(&html_path).map_or(0, |metadata| metadata.len());
-    println!(
+    writeln!(
+        output,
         "HTML report written to: {} ({} bytes)",
         html_path.display(),
         html_size
-    );
+    )?;
 
     // 5. JUnit-XML backend (CI ingestion).
     let junit_path = output_path.join("proof.xml");
-    report
-        .save(&JunitBackend, &junit_path)
-        .expect("JUnit-XML render failed");
+    report.save(&JunitBackend, &junit_path)?;
     let junit_size = std::fs::metadata(&junit_path).map_or(0, |metadata| metadata.len());
-    println!(
+    writeln!(
+        output,
         "JUnit-XML report written to: {} ({} bytes)",
         junit_path.display(),
         junit_size
-    );
+    )?;
 
-    println!(
+    writeln!(
+        output,
         "\n=== All five proof formats generated in {} ===",
         output_path.display()
-    );
+    )?;
+
+    Ok(())
 }
 
 /// Resolve the output directory for proof artifacts.
@@ -139,21 +146,31 @@ fn main() {
 /// Uses the first CLI argument if provided, otherwise defaults to
 /// `./testty_proof_output/` so generated files persist after the
 /// example exits.
-fn output_dir() -> PathBuf {
-    std::env::args()
-        .nth(1)
+fn output_dir(mut arguments: impl Iterator<Item = String>) -> PathBuf {
+    arguments
+        .next()
         .map_or_else(|| PathBuf::from("testty_proof_output"), PathBuf::from)
 }
 
 /// Print the first `max_lines` of a text file as a preview.
-fn print_file_preview(path: &Path, max_lines: usize) {
-    let content = std::fs::read_to_string(path).expect("failed to read file");
-    println!();
+fn print_file_preview(path: &Path, max_lines: usize, output: &mut impl Write) -> io::Result<()> {
+    let content = std::fs::read_to_string(path)?;
+    writeln!(output)?;
     for (index, line) in content.lines().enumerate() {
         if index >= max_lines {
-            println!("  ... ({} more lines)", content.lines().count() - max_lines);
+            writeln!(
+                output,
+                "  ... ({} more lines)",
+                content.lines().count() - max_lines
+            )?;
             break;
         }
-        println!("  {line}");
+        writeln!(output, "  {line}")?;
     }
+
+    Ok(())
 }
+
+#[cfg(test)]
+#[path = "proof_pipeline_test.rs"]
+mod tests;
